requirements.txt
djangorestframework
Django==2.1.5
setuptools==40.4.3